Made TestModelTemplate, TestModelTwoStep and TestModelBag into thermodynamics objects and removed TestModel file
